Реализован первый шаг нового аплоада с будущей балансировкой по нодам.
Первый шаг -- это получение ссылки для загрузки файла.
Ссылка теперь "получается" в редакторе карточек после рассчёта
контрольных сумм файла.

Новый Upload-контроллер на стороне сервера, теперь и будет заниматься
новым аплоадом, картинок или иных файлов.

SioController: вынос notifyWs() в соотв.модуль, занимающийся этим.

Реальной балансировки по нодам пока нет, но её нужно будет вставить
в эту первую фазу аплоада: дописывать в относительную ссылку адрес
хост для загрузки картинки. Надо просто перевести инфу от
weed assign в хостнейм ноды.

+++ src1/shared/common/src/main/scala/io/suggest/file/MSrvFileInfo.scala
@@ -20,7 +20,7 @@ object MSrvFileInfo {
   implicit val MSRV_FILE_INFO: OFormat[MSrvFileInfo] = (
     (__ \ "n").format[String] and
     (__ \ "u").format[String] and
-    (__ \ "s").formatNullable[Int] and
+    (__ \ "s").formatNullable[Long] and
     (__ \ "a").formatNullable[String] and
     (__ \ "m").formatNullable[String] and
     (__ \ "h").formatNullable[Map[MHash, String]]
@@ -47,7 +47,7 @@ object MSrvFileInfo {
 case class MSrvFileInfo(
                          nodeId     : String,
                          url        : String,
-                         sizeB      : Option[Int],
+                         sizeB      : Option[Long],
                          name       : Option[String],
                          mimeType   : Option[String],
                          hashesHex  : Map[MHash, String]
